fafo_projects/test_code_gen/generated_code/string_processing_pkg/2/agent_chat_pkg/tests/requirements.txt
# Test dependencies
pytest>=7.0.0
anthropic>=0.18.0
python-dotenv>=1.0.0
